Add sizes store, and put in a new admin view for sizes table data.
Also changed some files to add confirmation when the user tries to delete something.
Delete api.js as it is now not necessary with stores

+++ frontend/src/store/cobertura.js
@@ -33,6 +33,7 @@ const useCoberturaStore = defineStore('cobertura', {
 
         async delete_cobertura(cobertura) {
             try {
+                if (!confirm('Deseja realmente excluir essa cobertura?')) return;
                 await axiosClient.delete(`/api/coberturas/${cobertura.id}`);
                 this.coberturas = this.coberturas.filter(c => c.id !== cobertura.id);
             } catch (error) {
